simple-jobs-python-fastapi/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
tenacity>=8.2.0
